requirements.txt
pyyaml==6.0.1
schedule==1.2.0
sqlalchemy==2.0.25
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
psutil==5.9.8
cryptography==41.0.7
pytest==7.4.4
python-dotenv==1.0.0
aiohttp==3.9.1
prometheus-client==0.19.0
jinja2==3.1.3